Renamed steps around "merge" for better consistency.
TraveralMergeStep should have been MergeStep but that name was in use already as a base class for mergeV/E. Renamed that base class to MergeElementStep allowing merge() to use MergeStep.
